Game and Player now working

+++ SinkingShipsShip/src/main/java/at/fhv/sinkingshipsship/repositories/ShipRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ShipRepository extends JpaRepository<Ship, Integer> {
+public interface ShipRepository extends JpaRepository<Ship, Long> {
     List<Ship> findByGameId(Long gameId);
 }
